Added missing EndBlock asserts
