+++ Phantasma.Business/src/VM/Utils/DisasmUtils.cs
@@ -90,6 +90,7 @@ namespace Phantasma.Business.VM.Utils
             table["Runtime.InfuseToken"] = 5;
             table["Runtime.DeployContract"] = 4;
             table["Runtime.UpgradeContract"] = 4;
+            table["Runtime.KillContract"] = 2;
 
             table["Nexus.CreateToken"] = 3;
             table["Nexus.CreateTokenSeries"] = 7;
