+++ Phantasma.Tests/Exploits/Exploits.cs
@@ -15,40 +15,62 @@ namespace Phantasma.LegacyTests.Exploits;
 public class Exploits
 {
     [TestMethod]
-        public void ChainTransferExploit()
+    public void DuplicateTransferTest()
+    {
+        var owner = PhantasmaKeys.Generate();
+
+        var simulator = new NexusSimulator(owner);
+        var nexus = simulator.Nexus;
+
+        var target = PhantasmaKeys.Generate();
+
+        simulator.BeginBlock();
+        var tx1 = simulator.GenerateTransfer(owner, target.Address, simulator.Nexus.RootChain, DomainSettings.FuelTokenSymbol, 1);
+        var tx2 = simulator.GenerateTransfer(owner, target.Address, simulator.Nexus.RootChain, DomainSettings.FuelTokenSymbol, 1);
+
+        Assert.IsTrue(tx1.Hash == tx2.Hash);
+
+        simulator.EndBlock();
+        Assert.IsFalse(simulator.LastBlockWasSuccessful());
+    }
+
+    [TestMethod]
+    public void ChainTransferExploit()
+    {
+        var owner = PhantasmaKeys.FromWIF("L2LGgkZAdupN2ee8Rs6hpkc65zaGcLbxhbSDGq8oh6umUxxzeW25");
+
+        var simulator = new NexusSimulator(owner);
+        var nexus = simulator.Nexus;
+
+        var user = PhantasmaKeys.Generate();
+
+        var symbol = DomainSettings.StakingTokenSymbol;
+
+        var chainAddressStr = Base16.Encode(simulator.Nexus.RootChain.Address.ToByteArray());
+        var userAddressStr = Base16.Encode(user.Address.ToByteArray());
+
+        simulator.BeginBlock();
+        simulator.GenerateTransfer(owner, user.Address, simulator.Nexus.RootChain, DomainSettings.FuelTokenSymbol, 100000000);
+        simulator.GenerateTransfer(owner, user.Address, simulator.Nexus.RootChain, DomainSettings.StakingTokenSymbol, 100000000);
+        simulator.EndBlock();
+        Assert.IsTrue(simulator.LastBlockWasSuccessful());
+
+        var chainAddress = simulator.Nexus.RootChain.Address;
+        simulator.BeginBlock();
+        var tx = simulator.GenerateTransfer(owner, chainAddress, simulator.Nexus.RootChain, symbol, 100000000);
+        var block = simulator.EndBlock().FirstOrDefault();
+        Assert.IsTrue(simulator.LastBlockWasSuccessful());
+
+        var evts = block.GetEventsForTransaction(tx.Hash);
+        Assert.IsTrue(evts.Any(x => x.Kind == EventKind.TokenReceive && x.Address == chainAddress));
+
+        var token = simulator.Nexus.GetTokenInfo(simulator.Nexus.RootStorage, symbol);
+
+        var initialBalance = simulator.Nexus.RootChain.GetTokenBalance(simulator.Nexus.RootStorage, token, chainAddress);
+        Assert.IsTrue(initialBalance > 10000);
+
+        string[] scriptString = new string[]
         {
-            var owner = PhantasmaKeys.FromWIF("L2LGgkZAdupN2ee8Rs6hpkc65zaGcLbxhbSDGq8oh6umUxxzeW25");
-
-            var simulator = new NexusSimulator(owner);
-            var nexus = simulator.Nexus;
-
-            var user = PhantasmaKeys.Generate();
-
-            var symbol = DomainSettings.StakingTokenSymbol;
-
-            var chainAddressStr = Base16.Encode(simulator.Nexus.RootChain.Address.ToByteArray());
-            var userAddressStr = Base16.Encode(user.Address.ToByteArray());
-
-            simulator.BeginBlock();
-            simulator.GenerateTransfer(owner, user.Address, simulator.Nexus.RootChain, DomainSettings.FuelTokenSymbol, 100000000);
-            simulator.GenerateTransfer(owner, user.Address, simulator.Nexus.RootChain, DomainSettings.StakingTokenSymbol, 100000000);
-            simulator.EndBlock();
-
-            var chainAddress = simulator.Nexus.RootChain.Address;
-            simulator.BeginBlock();
-            var tx = simulator.GenerateTransfer(owner, chainAddress, simulator.Nexus.RootChain, symbol, 100000000);
-            var block = simulator.EndBlock().First();
-
-            var evts = block.GetEventsForTransaction(tx.Hash);
-            Assert.IsTrue(evts.Any(x => x.Kind == EventKind.TokenReceive && x.Address == chainAddress));
-
-            var token = simulator.Nexus.GetTokenInfo(simulator.Nexus.RootStorage, symbol);
-
-            var initialBalance = simulator.Nexus.RootChain.GetTokenBalance(simulator.Nexus.RootStorage, token, chainAddress);
-            Assert.IsTrue(initialBalance > 10000);
-
-            string[] scriptString = new string[]
-            {
                 $"alias r5, $sourceAddress",
                 $"alias r6, $targetAddress",
                 $"alias r7, $amount",
@@ -72,147 +94,148 @@ public class Exploits
                 $"push $targetAddress",
                 $"push $sourceAddress",
                 "extcall \"Runtime.TransferTokens\"",
-            };
+        };
 
-            var script = AssemblerUtils.BuildScript(scriptString);
+        var script = AssemblerUtils.BuildScript(scriptString);
 
-            simulator.BeginBlock();
-            simulator.GenerateCustomTransaction(user, ProofOfWork.None, () =>
-                ScriptUtils.BeginScript().
-                    AllowGas(user.Address, Address.Null, simulator.MinimumFee, Transaction.DefaultGasLimit).
-                    EmitRaw(script).
-                    SpendGas(user.Address).
-                    EndScript());
+        simulator.BeginBlock();
+        simulator.GenerateCustomTransaction(user, ProofOfWork.None, () =>
+            ScriptUtils.BeginScript().
+                AllowGas(user.Address, Address.Null, simulator.MinimumFee, Transaction.DefaultGasLimit).
+                EmitRaw(script).
+                SpendGas(user.Address).
+                EndScript());
 
-            try
-            {
-                simulator.EndBlock();
-            }
-            catch (Exception e)
-            {
-                Assert.IsTrue(e is ChainException);
-            }
+        simulator.EndBlock();
+        Assert.IsFalse(simulator.LastBlockWasSuccessful());
 
-            var finalBalance = simulator.Nexus.RootChain.GetTokenBalance(simulator.Nexus.RootStorage, token, simulator.Nexus.RootChain.Address);
-            Assert.IsTrue(initialBalance == finalBalance);
-        }
-        
-        [TestMethod]
-        public void TestPrintMoneyWithoutPermission()
+        var finalBalance = simulator.Nexus.RootChain.GetTokenBalance(simulator.Nexus.RootStorage, token, simulator.Nexus.RootChain.Address);
+        Assert.IsTrue(initialBalance == finalBalance);
+    }
+
+    [TestMethod]
+    public void TestPrintMoneyWithoutPermission()
+    {
+        var owner = PhantasmaKeys.FromWIF("L2LGgkZAdupN2ee8Rs6hpkc65zaGcLbxhbSDGq8oh6umUxxzeW25");
+
+        var simulator = new NexusSimulator(owner);
+        var nexus = simulator.Nexus;
+        nexus.SetOracleReader(new OracleSimulator(nexus));
+
+        var fuelToken = DomainSettings.FuelTokenSymbol;
+        var stakingToken = DomainSettings.StakingTokenSymbol;
+
+        //Let A be an address
+        var testUserA = PhantasmaKeys.Generate();
+
+        simulator.BeginBlock();
+        simulator.GenerateTransfer(owner, testUserA.Address, nexus.RootChain, fuelToken, 100000000);
+        simulator.GenerateTransfer(owner, testUserA.Address, nexus.RootChain, stakingToken, 100000000);
+        simulator.EndBlock();
+        Assert.IsTrue(simulator.LastBlockWasSuccessful());
+
+        var balanceBefore = nexus.RootChain.GetTokenBalance(simulator.Nexus.RootStorage, stakingToken, testUserA.Address);
+
+        simulator.BeginBlock();
+        simulator.GenerateCustomTransaction(testUserA, ProofOfWork.Minimal, () =>
+            ScriptUtils.BeginScript().AllowGas(testUserA.Address, Address.Null, 1, 9999)
+                .CallInterop("Runtime.MintTokens", "S3dP2jjf1jUG9nethZBWbnu9a6dFqB7KveTWU7znis6jpDy", testUserA.Address, stakingToken, 100000000).
+                SpendGas(testUserA.Address).EndScript());
+        simulator.EndBlock();
+        Assert.IsTrue(simulator.LastBlockWasSuccessful());
+
+        var balanceAfter = nexus.RootChain.GetTokenBalance(simulator.Nexus.RootStorage, stakingToken, testUserA.Address);
+        Assert.IsTrue(balanceAfter == balanceBefore);
+        Assert.IsFalse(balanceAfter == balanceBefore + 100000000);
+    }
+
+    public struct crownStruct
+    {
+        public Address staker;
+        public Timestamp date;
+    }
+
+    [TestMethod]
+    public void TestPrintCrowns()
+    {
+        var owner = PhantasmaKeys.FromWIF("L2LGgkZAdupN2ee8Rs6hpkc65zaGcLbxhbSDGq8oh6umUxxzeW25");
+
+        var simulator = new NexusSimulator(owner);
+        var nexus = simulator.Nexus;
+        nexus.SetOracleReader(new OracleSimulator(nexus));
+
+        var fuelToken = DomainSettings.FuelTokenSymbol;
+        var stakingToken = DomainSettings.StakingTokenSymbol;
+
+        //Let A be an address
+        var testUserA = PhantasmaKeys.Generate();
+
+        var crownRom = new crownStruct
         {
-            var owner = PhantasmaKeys.FromWIF("L2LGgkZAdupN2ee8Rs6hpkc65zaGcLbxhbSDGq8oh6umUxxzeW25");
+            staker = testUserA.Address,
+            date = Timestamp.Now
+        };
 
-            var simulator = new NexusSimulator(owner);
-            var nexus = simulator.Nexus;
-            nexus.SetOracleReader(new OracleSimulator(nexus));
+        var crownRam = 1;
 
-            var fuelToken = DomainSettings.FuelTokenSymbol;
-            var stakingToken = DomainSettings.StakingTokenSymbol;
+        simulator.BeginBlock();
+        simulator.GenerateTransfer(owner, testUserA.Address, nexus.RootChain, fuelToken, 100000000);
+        simulator.GenerateTransfer(owner, testUserA.Address, nexus.RootChain, stakingToken, 100000000);
+        simulator.EndBlock();
+        Assert.IsTrue(simulator.LastBlockWasSuccessful());
 
-            //Let A be an address
-            var testUserA = PhantasmaKeys.Generate();
+        var reward = new StakeReward(testUserA.Address, Timestamp.Now);
 
-            simulator.BeginBlock();
-            simulator.GenerateTransfer(owner, testUserA.Address, nexus.RootChain, fuelToken, 100000000);
-            simulator.GenerateTransfer(owner, testUserA.Address, nexus.RootChain, stakingToken, 100000000);
-            simulator.EndBlock();
+        var balanceBefore = nexus.RootChain.GetTokenBalance(simulator.Nexus.RootStorage, "CROWN", testUserA.Address);
 
-            var balanceBefore = nexus.RootChain.GetTokenBalance(simulator.Nexus.RootStorage, stakingToken, testUserA.Address);
-            
-            simulator.BeginBlock();
-            simulator.GenerateCustomTransaction(testUserA, ProofOfWork.Minimal, () =>
-                ScriptUtils.BeginScript().AllowGas(testUserA.Address, Address.Null, 1, 9999)
-                    .CallInterop("Runtime.MintTokens", "S3dP2jjf1jUG9nethZBWbnu9a6dFqB7KveTWU7znis6jpDy", testUserA.Address, stakingToken, 100000000).
-                    SpendGas(testUserA.Address).EndScript());
-            simulator.EndBlock();
 
-            var balanceAfter = nexus.RootChain.GetTokenBalance(simulator.Nexus.RootStorage, stakingToken, testUserA.Address);
-            Assert.IsTrue(balanceAfter == balanceBefore );
-            Assert.IsFalse(balanceAfter == balanceBefore + 100000000);
-        }
-        
-        public struct crownStruct 
-        {
-            public Address staker;
-            public Timestamp date;
-        }
-        
-        [TestMethod]
-        public void TestPrintCrowns()
-        {
-            var owner = PhantasmaKeys.FromWIF("L2LGgkZAdupN2ee8Rs6hpkc65zaGcLbxhbSDGq8oh6umUxxzeW25");
+        simulator.BeginBlock();
+        simulator.GenerateCustomTransaction(testUserA, ProofOfWork.Minimal, () =>
+            ScriptUtils.BeginScript().AllowGas(testUserA.Address, Address.Null, 1, 9999)
+                .CallInterop("Runtime.MintToken", "S3d79FvexQeerRioAY3pGYpNPFx7oJkMV4KazdTHdGDA5iy", testUserA.Address, DomainSettings.RewardTokenSymbol, VMObject.FromStruct(reward), new byte[0], 0).
+                SpendGas(testUserA.Address).EndScript());
+        simulator.EndBlock();
+        Assert.IsTrue(simulator.LastBlockWasSuccessful());
 
-            var simulator = new NexusSimulator(owner);
-            var nexus = simulator.Nexus;
-            nexus.SetOracleReader(new OracleSimulator(nexus));
+        var balanceAfter = nexus.RootChain.GetTokenBalance(simulator.Nexus.RootStorage, "CROWN", testUserA.Address);
+        Assert.IsTrue(balanceAfter == balanceBefore, $"{balanceAfter} != {balanceBefore}");
+    }
 
-            var fuelToken = DomainSettings.FuelTokenSymbol;
-            var stakingToken = DomainSettings.StakingTokenSymbol;
+    [TestMethod]
+    public void TestInfiniteTokenTransfer()
+    {
+        var owner = PhantasmaKeys.Generate();
 
-            //Let A be an address
-            var testUserA = PhantasmaKeys.Generate();
+        var simulator = new NexusSimulator(owner);
+        var nexus = simulator.Nexus;
 
-            var crownRom = new crownStruct
-            {
-                staker = testUserA.Address,
-                date = Timestamp.Now
-            };
+        simulator.BeginBlock();
+        simulator.GenerateToken(owner, "INFI", "infinity token", 0, 8,
+            TokenFlags.Fungible | TokenFlags.Transferable | TokenFlags.Divisible);
+        simulator.EndBlock();
+        Assert.IsTrue(simulator.LastBlockWasSuccessful());
 
-            var crownRam = 1;
+        var user = PhantasmaKeys.Generate();
 
-            simulator.BeginBlock();
-            simulator.GenerateTransfer(owner, testUserA.Address, nexus.RootChain, fuelToken, 100000000);
-            simulator.GenerateTransfer(owner, testUserA.Address, nexus.RootChain, stakingToken, 100000000);
-            simulator.EndBlock();
-            
-            var reward = new StakeReward(testUserA.Address, Timestamp.Now);
-            
-            var balanceBefore = nexus.RootChain.GetTokenBalance(simulator.Nexus.RootStorage, "CROWN", testUserA.Address);
+        var infiToken = nexus.GetTokenInfo(nexus.RootStorage, "INFI");
 
-            
-            simulator.BeginBlock();
-            simulator.GenerateCustomTransaction(testUserA, ProofOfWork.Minimal, () =>
-                ScriptUtils.BeginScript().AllowGas(testUserA.Address, Address.Null, 1, 9999)
-                    .CallInterop("Runtime.MintToken", "S3d79FvexQeerRioAY3pGYpNPFx7oJkMV4KazdTHdGDA5iy", testUserA.Address, DomainSettings.RewardTokenSymbol, VMObject.FromStruct(reward),  new byte[0], 0).
-                    SpendGas(testUserA.Address).EndScript());
-            simulator.EndBlock();
+        var infiAmount = 1000 * UnitConversion.GetUnitValue(infiToken.Decimals);
 
-            var balanceAfter = nexus.RootChain.GetTokenBalance(simulator.Nexus.RootStorage, "CROWN", testUserA.Address);
-            Assert.IsTrue(balanceAfter == balanceBefore, $"{balanceAfter} != {balanceBefore}");
-        }
-        
-        [TestMethod]
-        public void TestInfiniteTokenTransfer()
-        {
-            var owner = PhantasmaKeys.Generate();
+        simulator.BeginBlock();
+        simulator.GenerateTransfer(owner, user.Address, nexus.RootChain, DomainSettings.FuelTokenSymbol, 100000000);
+        simulator.MintTokens(owner, owner.Address, infiToken.Symbol, infiAmount);
+        simulator.EndBlock();
+        Assert.IsTrue(simulator.LastBlockWasSuccessful());
 
-            var simulator = new NexusSimulator(owner);
-            var nexus = simulator.Nexus;
+        var balance = nexus.RootChain.GetTokenBalance(simulator.Nexus.RootStorage, infiToken, owner.Address);
+        Assert.IsTrue(balance == infiAmount);
 
-            simulator.BeginBlock();
-            simulator.GenerateToken(owner, "INFI", "infinity token", 0, 8,
-                TokenFlags.Fungible | TokenFlags.Transferable | TokenFlags.Divisible);
-            simulator.EndBlock();
+        simulator.BeginBlock();
+        simulator.GenerateTransfer(owner, user.Address, nexus.RootChain, infiToken.Symbol, infiAmount);
+        simulator.EndBlock();
+        Assert.IsTrue(simulator.LastBlockWasSuccessful());
 
-            var user = PhantasmaKeys.Generate();
-
-            var infiToken = nexus.GetTokenInfo(nexus.RootStorage, "INFI");
-
-            var infiAmount = 1000 * UnitConversion.GetUnitValue(infiToken.Decimals);
-
-            simulator.BeginBlock();
-            simulator.GenerateTransfer(owner, user.Address, nexus.RootChain, DomainSettings.FuelTokenSymbol, 100000000);
-            simulator.MintTokens(owner, owner.Address, infiToken.Symbol, infiAmount);
-            simulator.EndBlock();
-
-            var balance = nexus.RootChain.GetTokenBalance(simulator.Nexus.RootStorage, infiToken, owner.Address);
-            Assert.IsTrue(balance == infiAmount);
-
-            simulator.BeginBlock();
-            simulator.GenerateTransfer(owner, user.Address, nexus.RootChain, infiToken.Symbol, infiAmount);
-            simulator.EndBlock();
-
-            balance = nexus.RootChain.GetTokenBalance(simulator.Nexus.RootStorage, infiToken, user.Address);
-            Assert.IsTrue(balance == infiAmount);
-        }
+        balance = nexus.RootChain.GetTokenBalance(simulator.Nexus.RootStorage, infiToken, user.Address);
+        Assert.IsTrue(balance == infiAmount);
+    }
 }
