+++ Phantasma.Tests/Exploits/WhenSystemAddressUsedAsSourceInScript.cs
@@ -55,8 +55,9 @@ public class WhenSystemAddressUsedAsSourceInScript
         simulator.GenerateTransfer(owner, address, nexus.RootChain, DomainSettings.FuelTokenSymbol, initialFuel);
         simulator.GenerateTransfer(owner, address, nexus.RootChain, DomainSettings.StakingTokenSymbol, initialAmount);
         simulator.EndBlock();
+        Assert.IsTrue(simulator.LastBlockWasSuccessful());
     }
-          
+
     [TestMethod("Mint Token Should Not Mint Token.")]
     public void MintTokenShouldNotMintToken()
     {
@@ -68,7 +69,7 @@ public class WhenSystemAddressUsedAsSourceInScript
                 .CallInterop("Runtime.MintToken", sysAddress, user.Address, DomainSettings.RewardTokenSymbol, VMObject.FromStruct(reward), new byte[0], 0).
                 SpendGas(user.Address).EndScript());
         simulator.EndBlock();
-        //Assert.ThrowsException<ChainException>(() => simulator.EndBlock());
+        Assert.IsFalse(simulator.LastBlockWasSuccessful());
 
         //Assertions if exception is not used in this circumstance
         var balanceAfter = nexus.RootChain.GetTokenBalance(simulator.Nexus.RootStorage, DomainSettings.RewardTokenSymbol, user.Address);
@@ -84,8 +85,7 @@ public class WhenSystemAddressUsedAsSourceInScript
             ScriptUtils.BeginScript().AllowGas(user.Address, Address.Null, 1, gas)
                 .CallInterop("Runtime.MintTokens", sysAddress, user.Address, DomainSettings.StakingTokenSymbol, amountRequested).
                 SpendGas(user.Address).EndScript());
-        simulator.EndBlock();
-        //Assert.ThrowsException<ChainException>(() => simulator.EndBlock());
+        Assert.IsFalse(simulator.LastBlockWasSuccessful());
 
         var endBalance = nexus.RootChain.GetTokenBalance(simulator.Nexus.RootStorage, DomainSettings.StakingTokenSymbol, user.Address);
         Assert.AreNotEqual(startBalance + amountRequested, endBalance, "Requested tokens were minted");
@@ -103,8 +103,7 @@ public class WhenSystemAddressUsedAsSourceInScript
                 .CallInterop("Runtime.MintTokens", owner.Address, sysAddress, DomainSettings.StakingTokenSymbol, amountRequested)
                 .CallInterop("Runtime.TransferTokens", sysAddress, user.Address, DomainSettings.StakingTokenSymbol, amountRequested).
                 SpendGas(user.Address).EndScript());
-        simulator.EndBlock();
-        //Assert.ThrowsException<ChainException>(() => simulator.EndBlock());
+        Assert.IsFalse(simulator.LastBlockWasSuccessful());
 
         var endBalance = nexus.RootChain.GetTokenBalance(simulator.Nexus.RootStorage, DomainSettings.StakingTokenSymbol, user.Address);
         Assert.AreNotEqual(startBalance + amountRequested, endBalance, "Requested tokens were transfered");
